More secure way of storing the Todoist API key
It's considered bad practice to store your todoist API key out in the
open. You could accidentally share your bashrc and give someone control
to your Todoist account. Pass is a software which encrypts password and
make them easily accessible. This way, your API key is never stored out
in the open and your key is as easily accessible to the plugin.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   README.md